backend/chatbot/requirements.txt
# FastAPI core
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database (PostgreSQL + ORM)
sqlalchemy==2.0.35
psycopg2-binary==2.9.10   # PostgreSQL driver
alembic==1.13.2          # DB migrations

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT handling
passlib[bcrypt]==1.7.4            # Password hashing
python-multipart==0.0.9           # Form-data parsing (file uploads)

# Environment variables
python-dotenv==1.0.1

# API Docs & Validation
pydantic>=2.7,<3.0
pydantic-settings>=2.0,<3.0

# Redis (for session/chat history)
redis==5.1.0

# CORS & Middleware
starlette==0.38.5

# Testing
httpx==0.27.0
pytest==8.3.3


aiofiles==23.2.1
requests==2.31.0
python-jwt==4.0.0

PyJWT
